i18n: using msgmerge instead of pybabel update
The algorithm used by "pybabel update" to detect small changes that do
not require a new translation but an update of an existing
translation (also called fuzzy updates) is very different from the
algorithm used by msgmerge. See the following babel issue for more
information:

  https://github.com/python-babel/babel/issues/524

It turns out that the "pybabel update" algorithm does not work so well
with the SecureDrop strings: a string ending with an exclamation mark
instead of a dot will not be found to be similar.

We using msgmerge instead to get more fuzzy matches and ease the work
of translators. The cost of not detecting a similar translation is
high because the translator has to refer to the git history to figure
out how it was translated previously. git tool is not familiar to most
translators.